requirements.txt
pyyaml>=6.0
pydantic>=2.0
pydub>=0.25.1
librosa>=0.10.0
numpy>=1.24
requests>=2.28
